x/tictactoe/types/message_game_join_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"tictactoe/testutil/sample"
)

func TestMsgGameJoin_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgGameJoin
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgGameJoin{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgGameJoin{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
